Added listener commands, dropped CLI dependency
IOFMessageListener can now return a Command enum from the receive
method that will control further calls to listeners for that particular
message.  Added a test to verify.

Listeners also can now throw exceptions which will be caught by
the controller and not propagated.

Added further testing to insure ordering is correct when specified.