Topology and DeviceManager threading and object changes
Changed the structures in Topology and DeviceManager to
index a switch by the actual IOFSwitch object instead
of the Long id.

Changed Topology to auto wire itself with all
published services exporting the TopologyAware interface.

Changed Topology to provide updates to TopologyAware
objects in a guaranteed ordering as processed by Topology
itself.  These updates are passed to listeners by a single
thread.

Added code to DeviceManager to remove devices who's location
is an internal port once it is discovered via Topology that
this is the case.