Removed interest-op setting from the Streams
This cannot work correctly because setting
the key ops must occur from within the thread
handling the I/O and disconnect status of
the switch, otherwise cancelled key exceptions
occur, or heavy locking must be introduced
to protect the keys.

Further currently it really does nothing,
so no reason to have it.